+++ UM_LOGISTIC_V1/Client/Controllers/Transportation/TransportationController.js
@@ -41,7 +41,7 @@
 		.success(response => {
 		    $scope.isLoading = false;
 		    $scope.isPartLoading = false;
-		    if (response.Success) {
+		    if (response && response.Success == true) {
 		        for (let i = 0; i < response.Result.length; i++) {
 		            $scope.transportations.push({
 		                id: FormHelper.getFormValue(response.Result[i].Id),
@@ -70,7 +70,7 @@
 		    $scope.isPartLoading = false;
 		    $scope.isLoading = false;
 		    $scope.filterLoading = false;
-		    if (response.Success) {
+		    if (response && response.Success == true) {
 		        if (page == 0) {
 		            $scope.transportations = [];
 		        }
@@ -111,7 +111,7 @@
         let type = true;
         ApplicationPictureService.getApplicationPictures(id, type)
 		.success(response => {
-		    if (response.Success && response.Result && response.Result.length > 0) {
+		    if (response && response.Success == true && response.Result && response.Result.length > 0) {
 		        $scope.pictures[id] = response.Result[0];
 		    }
 		    else {
@@ -137,7 +137,7 @@
         request.TypeId = 2;
         ClientTaskService.createApplicationTask(request)
         .success(response => {
-            if (response.Success) {
+            if (response && response.Success == true) {
                 NotificationService.success(moduleConstants.callFeedbackAccepted);
                 $scope.isLoading = false;
             }
@@ -156,7 +156,6 @@
         let filter = $scope.filter;
         let stringFilter = $scope.generateFilterString(filter);
         $scope.currentPage = 0;
-        //$scope.isLoading = true;
         $scope.filterLoading = true;
         $scope.listFilteredTransportations(stringFilter, $scope.currentPage, $scope.currentCount);
     }
@@ -272,10 +271,10 @@
             UserId: userId
         };
         ApplicationTrashService.createApplicationTrash(request).success(response => {
-            if (response.Success == false) {
+            if (response && response.Success == false) {
                 NotificationService.error(response.Error != null ? JSON.stringify(response.Error) : moduleConstants.internalErrorCaption);
             }
-            if (response.Success == true) {
+            if (response && response.Success == true) {
                 let isTrashAdded = SessionService.addTrashElement(id, true, title);
                 if (isTrashAdded == true) {
                     $rootScope.$broadcast("trashElementAdded", null);
@@ -301,7 +300,7 @@
         }
         ApplicationTrashService.removeTrashElement(id, true)
         .success(response => {
-            if (response.Success) {
+            if (response && response.Success == true) {
                 $scope.isLoading = false;
                 SessionService.removeShopTrashElement(id, true);
                 $rootScope.$broadcast("trashElementRemoved", null);
